Simplify regex lexer
Instead of proactively joining tokens of the same type, it now just
collects tokens without caring about that. This heavily simplifies the
process, because instead of collecting the previous match when the next
type is different, the match is instantly collected.

This makes the process a lot easier to reason about, because only one
position has to be kept track of.

To merge tokens of the same type together, a filter could be added just
like Pygments has in place. Infrastructure for this doesn't yet exist.