Add testing framework and comparison to Pygments
The system CRYLIC-TEST forms the basis for testing and will contains
standalone tests for Crylic. I decided to go with 1am because it's
lightweight, simple and easy to use.

Because some tests should be separate from others, I implemented a
little layer over 1am to separate some tests from others, codified by
the macro DEFINE-TEST-FRAMEWORK.

The system CRYLIC-TEST-VS-PYGMENTS depends on CRYLIC-TEST and exists to
compare the output of Crylic to Pygments. It is separate because normal
tests should not depend on `pygmentize` being installed, and
compatibility with Pygments should be optional, not a requirement for
tests to succeed.

Do note that the example files, which are currently only used by the
CRYLIC-TEST-VS-PYGMENTS system, are specified as part of the CRYLIC-TEST
system. This is because they won't only be useful for comparison, but
also for general testing.