Use modeless start anchor (\A) for every regex
Every regex should only ever match at the current start of input.
Previously, a start anchor (^) was prepended to every match. This
matched either at the start of input or at a new line. To eliminate
false matches, when a regex matched the match index was checked to make
sure it was actually at the right location. This meant CL-PPCRE
sometimes did a lot more work than was really necessary.

Replacing the start anchor (^) with a modeless start anchor (\A) that
only ever matches at the start of input (and thus also eliminates the
extra check) improves performance by quite a bit.